Move evaluators that count by window to rulepassed
